chatbot-vocacional-main/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
distro==1.9.0
ecdsa==0.19.0
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.7.0/es_core_news_md-3.7.0-py3-none-any.whl#sha256=0d6d6ebed875869a9759c8c096f2cef581fa32d861646030f771c83e5799de82
fastapi==0.110.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
ollama==0.2.1
openai==1.31.0
packaging==24.0
passlib==1.7.4
preshed==3.0.9
psycopg2==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==69.0.3
sib-api-v3-sdk==7.6.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
thinc==8.2.3
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.10.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
wheel==0.43.0
